Snapcast u pn p (#1)
* bug fixess, getops and module additions

- each function is now callable on an individual level with command line flags
- if statement for debian version checks.
- inclusion of snapcast installation
- inclusion of UPnP client installation
- README updated to include new modules.

new version of shairport and settings defaults changed to be a bit more useful and a little less deafening.

* Update install.sh

fix getops arguments and include fix lost with merge

* Update install.sh

updated to have clearer communication of progress (banners and colouring)

Added a few 'basic os' setup commands. Specifically netselect-apt

the function update_latest is not set to execute. I'm not sure it's a good idea to force it on people just yet.

* Update install.sh

- heading text returns to normal colour after writing a message.
- Fix mistakes with function and variable calls.
- enabling bt-agent '--now' flag removed due to crashes
- updated Get Opts help message.
- moved OS verification and netselect to the standard installation path. With yes no flag.